remove too strict lifetime binding
the lifetime in SearchQueryOperation were binding external lifetime to
self, and this was making non trivial use of the SearchQueryOperation
very difficult.
as you can see in a minimal setup
https://play.rust-lang.org/?version=stable&mode=debug&edition=2015&gist=ca2db06f73dbcebdfbd7049232b727ad

this was leading to multiple mutable borrow if we want to keep a
reference on the search_query.

fixed version:
https://play.rust-lang.org/?version=stable&mode=debug&edition=2015&gist=25482710b5ebc7f1dded1a33fcf7c8c0

Note: I had the problem when trying to set an optional timeout, as you
can see
[here](https://github.com/antoine-de/mimirsbrunn/blob/poc_borrow/libs/bragi/src/query.rs#L272-L284),
the workaround was a bit uggly.

thanks @TeXitoi for the huge help in understanding the problem!